feat (core): added account deactivation to telegram service, updated cron script to flush of deactivated accounts and updated webhook to notify of users

+++ src/modules/cron/cron.service.ts
@@ -3,13 +3,15 @@ import { Injectable } from "@nestjs/common";
 import { MailService } from "../libs/mail/mail.service";
 import { Cron } from "@nestjs/schedule";
 import { StorageService } from "../libs/storage/storage.service";
+import { TelegramService } from "../libs/telegram/telegram.service";
 
 @Injectable()
 export class CronService {
 	public constructor(
 		private readonly prismaService: PrismaService,
 		private readonly mailService: MailService,
-		private readonly storageService: StorageService
+		private readonly storageService: StorageService,
+		private readonly telegramService: TelegramService
 	) {}
 
 	@Cron("0 0 * * *")
@@ -24,12 +26,21 @@ export class CronService {
 					lte: sevenDaysAgo,
 				},
 			},
+			include: {
+				notificationSettings: true,
+				stream: true
+			},
 		});
 
 		for (const user of deactivatedAccounts) {
 			await this.mailService.sendAccoundDeletion(user.email);
 
-			this.storageService.remove(user.avatar as string);
+			if (user.notificationSettings && user.telegramId) {
+				await this.telegramService.sendAccountDeletion(user.telegramId);
+			}
+
+			if (user.avatar) this.storageService.remove(user.avatar as string);
+			if (user.stream?.thumbnailUrl) this.storageService.remove(user.stream?.thumbnailUrl as string);
 		}
 
 		await this.prismaService.user.deleteMany({
